+++ src/test/java/ua/foxminded/moldavets/project/storage/FileStorageImplTest.java
@@ -5,6 +5,8 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import ua.foxminded.moldavets.project.exception.ExistStorageException;
+import ua.foxminded.moldavets.project.exception.NotExistStorageException;
+import ua.foxminded.moldavets.project.exception.StorageException;
 import ua.foxminded.moldavets.project.model.Resume;
 import ua.foxminded.moldavets.project.storage.serializer.ObjectStreamSerializer;
 
@@ -34,7 +36,9 @@ class FileStorageImplTest {
     @AfterEach
     void tearDown() {
         File toBeDeleted = new File("C:\\Users\\steam\\IdeaProjects\\Java-SE-app\\storage\\testStorage");
-        deleteDirectory(toBeDeleted);
+        if (toBeDeleted.listFiles() != null) {
+            deleteDirectory(toBeDeleted);
+        }
     }
 
     @Test
@@ -56,6 +60,89 @@ class FileStorageImplTest {
         Exception actual = assertThrows(NullPointerException.class,
                 () -> fileStorage.save(resume1, null));
         assertEquals("Resume or search key cannot be null", actual.getMessage());
+    }
+
+    @Test
+    void save_shouldReturnException_whenInputSearchKeyAndResumeIsEmpty() {
+        File file = new File("");
+        Resume resume = new Resume("","");
+        Exception actual = assertThrows(StorageException.class,
+                () -> fileStorage.save(resume, file));
+        assertEquals("Couldn't save resume " + resume.getUuid() + "to " +
+                file.getAbsolutePath() + "\\" + resume.getUuid(),
+                actual.getMessage());
+    }
+
+    @Test
+    void update_shouldReturnException_whenInputResumeDoesNotExist() {
+        Resume resume = new Resume("testUuid1", "testFullName");
+        Exception actual = assertThrows(NotExistStorageException.class,
+                () -> fileStorage.update(resume, fileStorage.getSearchKey(resume.getUuid())));
+        assertEquals("Resume " + resume.getUuid() + " does not exist", actual.getMessage());
+    }
+
+    @Test
+    void update_shouldUpdateResume_whenInputResumeExists() {
+        Resume resume4 = new Resume("uuid1","Shara Doe");
+        fileStorage.update(resume4, fileStorage.getSearchKey(resume4.getUuid()));
+
+        assertEquals("Shara Doe", fileStorage.get(fileStorage.getSearchKey(resume1.getUuid())).getFullName());
+    }
+
+    @Test
+    void update_shouldReturnException_whenInputSearchKeyIsNull() {
+        Exception actual = assertThrows(NullPointerException.class,
+                () -> fileStorage.update(resume1, null));
+        assertEquals("Resume or search key cannot be null", actual.getMessage());
+    }
+
+    @Test
+    void update_shouldReturnException_whenInputResumeIsNull() {
+        Exception actual = assertThrows(NullPointerException.class,
+                () -> fileStorage.save(null, null));
+        assertEquals("Resume or search key cannot be null", actual.getMessage());
+    }
+
+    @Test
+    void delete_shouldReturnException_whenInputResumeDoesNotExist() {
+        Resume resume = new Resume("testUuid1", "testFullName");
+        Exception actual = assertThrows(StorageException.class,
+                () -> fileStorage.delete(fileStorage.getSearchKey(resume.getUuid())));
+        assertEquals("File delete exception", actual.getMessage());
+    }
+
+    @Test
+    void delete_shouldDeleteResume_whenInputResumeExists() {
+        fileStorage.delete(fileStorage.getSearchKey(resume1.getUuid()));
+        assertFalse(fileStorage.isExist(fileStorage.getSearchKey(resume1.getUuid())));
+    }
+
+    @Test
+    void delete_shouldReturnException_whenInputSearchKeyIsNull() {
+        Exception actual = assertThrows(NullPointerException.class,
+                () -> fileStorage.delete( null));
+        assertEquals("File cannot be null", actual.getMessage());
+    }
+
+    @Test
+    void delete_shouldReturnException_whenInputSearchKeyIsEmpty() {
+        Exception actual = assertThrows(StorageException.class,
+                () -> fileStorage.delete(new File("")));
+        assertEquals("File delete exception", actual.getMessage());
+    }
+
+    @Test
+    void clear_shouldReturnException_whenStorageIsEmpty() {
+        deleteDirectory(new File("C:\\Users\\steam\\IdeaProjects\\Java-SE-app\\storage\\testStorage"));
+        Exception actual = assertThrows(StorageException.class,
+                fileStorage::clear);
+        assertEquals("Storage is empty", actual.getMessage());
+    }
+
+    @Test
+    void clear_shouldClearStorage_whenStorageIsNotEmpty() {
+        fileStorage.clear();
+        assertEquals(directory.listFiles().length, 0);
     }
